Configure message processor for max retries
This changes the message processor to use the new Message class where
there is just done and retry as options.

It then configures the places where there are retries to quit after
exhausting max retries.

While we could handle the max retries as part of a method on the Message
object, this was rejected as it would lead to confusing logging.