Update OpenAIStructuredAnswerComposer to abort if LLM cannot answer
This change updates the OpenAIStructuredAnswerComposer to abort if the
LLM cannot answer the question. When the LLM provides the structured
response, one of the key value pairs in the JSON response is "answered".

If the value is false then we know that the LLM could not answer the
question and we can abort the pipeline and assign the answer the
abort_llm_cannot_answer status.

For now i've updated the message to the "Sorry, I am not able to answer
your question." since that is our suggested response in the prompt.

We will want to chat about this with Content Design to check that this is
okay.

Once we've added question routing that should handle most of these cases
so something generic like the above should be fine as it shouldn't happen
often.