Add TODO comment for adding non-null constraints
We can't add these constraints until we remove the `govuk_chat_api`
answer strategy as it doesn't pull content from OpenSearch and so we
don't have values for the chunk IDs or digests.

Once we remove it, all answer sources will be linked to a chunk in the
content store, so we'd want to make sure these values are always
present.