Change the way buttons are disabled while awaiting a response
When the form button is disabled during question submission, focus defaults to the body element, interrupting announcements in the conversation list container. The changes ensure that focus is maintained while the button appears visually disabled.

The `aria-disabled` attribute is used to prevent double-clicking on the button, avoiding temporary loss of keyboard focus during form submission, an issue that occurs with the disabled attribute.

Also, visually hidden text in the button component is updated to indicate that a response is awaited from GOV.UK Chat, explaining the 'disabled' state.