Add used boolean to answer sources table
Now we can determine which sources the LLM used, we need a way to indicate
which sources were used so we can ensure the user is only shown the sources
used to generate the answer.

This commit adds a 'used' boolean column to the answer_sources table to
store this information.