Replace div chunking with HTML flattening
This changes our approach to chunking to bring it closer to the
implementation made in the data science prototype [1].

Previously we treated a div as the start of a new chunk and now we
instead flatten div elements so they can be part of an existing chunk.
We also expanded the list of elements we flatten to cover other block
level structural elements. I'm not sure whether we'll see any of these
other than div in our input, but it seemed better to be prepared for an
odd situation than it was to have a bad document.

The elements flattened aren't exactly the same as the prototype [2]. I
removed svg and path as we shouldn't be expecting to find page content
in a SVG and if we did want to we'd have 10s of elements to consider. I
removed span as it would be semantically incorrect to have headers
inside a span and thus shouldn't happen.

[1]: https://github.com/alphagov/govuk-vectorstore-data-pipeline/blob/html-hierarchical-chunking/html_hierarchical_chunking/html_splitter_beautifulsoup.py