Update sources based on the sources used by LLM
This adds an #update_sources_from_exact_paths_used method to the context
model. This method takes an array of exact_paths for sources the LLM
used to generate the answer.

It iterates through the exact paths. For each exact path, it finds the
source on the answer and sets used to true and sets updates the relevancy
to the index of the exact_path.

If no exact_paths are passed in then it marks all sources as used and
returns.

If it is unable to find a source for an exact path, it ignores them by
compacting the array to remove nils.

It then sets all unused sources 'used' attribute to false and updates the
relevancies to be higher than the highest relevancy of the used sources.