Set all sources to unused when abort_pipeline is called
For now, there are no situations when we abort the pipeline where we
want to show the user any sources.

This commit sets all sources to unused when abort_pipeline is called.

It might be down the line that we want to show the users sources in
some situations. If this ever is the case we could update abort_pipeline
to make an argument for sources, but for now there is no need.