Update #group_answer_sources_by_base_path to filter out unused sources
Now that the abort_pipeline method has been updated to update all
an answers sources used to "false" when the pipeline is aborted, we
can be confident that if a source has used "true", the answer has a
status of "success".

This commit add a ''.used' scope, updates the #group_answer_sources_by_base_path
method to to #group_used_answer_sources_by_base_path and applies the scope
to filter out the unused sources.