Refactor admin search results for reduced coupling
This resolves a failing test which occurs because the close embedding is
no longer a predictable number now that the embedding dimensions have
changed.

Along the way of fixing these I wanted to reduce the tightness of the
coupling between this test and changes to application behaviour so I've
reduced the number of assertions and fidelity of them.

I've also added id's for each table to be able to distinguish them
easier than using nth element. I wanted to add the id to the table but
the component doesn't allow that :( [1]

[1]: https://github.com/alphagov/govuk_publishing_components/blob/840d1cb28ea7f05e60bb18ef26a9cce66d847967/app/views/govuk_publishing_components/components/_table.html.erb#L13