Switch from l2 space_type to cosine similarity
This changes the distance functions for approximate nearest neighbour
searches to match the cosine similarity function that was used for the
prototype. This has been done to improve the accuracy of our results.

A consequence of this change is that we no longer get as predictable
scores when we do a comparison of exact embeddings and now get a 0.9999
esque number rather than a rounded 1, thus we have to change a number of
tests.