Rename url field to exact_path
This completes a todo left for the next time we reindex chunked content
to rename the url field to be exact_path.

The motivation for this renaming is to reflect the renaming that took
place in sources where we have base_path and exact_path. A exact_path
reflects a more accurate path to a piece of GOV.UK content and may
include a fragment.

This change required a lot of shotgun surgery :-(