Retry all unknown exceptions in MessageProcessor
Now that we have a delayed retry system that will only retry a few times
it seems of low negative consequence to apply this to all the exceptions
we receive and thus increase the chance of automatically resolving a
transient error.

This change applies the same retry approach we were using for our known
retry errors, OpenSearch::Transport::Transport::Error and
OpenAIClient::RequestError, to StandardError. It has an exception for
when a payload is not a hash as that seems uniformly terminal.