Use basic dynos for review apps
The default eco dynos we use are on a limited amount of hours per month
and we as an organisation are prone to exceeding the limit.

For branches to make use of this we'll need them to rebase this branch,
delete their existing review app and then create a new one.