Parts parser: ignore slug in URL for the first part
When we parse a content item for a "part" schema (e.g. travel advice),
we loop over each part and index that as a discrete chunk.

Each chunk has an `exact_path` which we're currently building by taking
the base path and appending the part's slug to it.

This works fine for all parts, but on the actual GOV.UK website, the
first part doesn't include the slug.

For example, take the travel advice page for Afghanistan[1]. It has a
table of contents listing each part:

- Warnings and insurance
- Entry requirements
- Safety and security
...

The URL for the first item in the list is just the base URL, it doesn't
have a slug. So the URLs look like this:

- Warnings and insurance: https://www.gov.uk/foreign-travel-advice/afghanistan
- Entry requirements: https://www.gov.uk/foreign-travel-advice/afghanistan/entry-requirements
- Safety and security: https://www.gov.uk/foreign-travel-advice/afghanistan/safety-and-security

Technically the first part _does_ have a slug which is present in the
payload we're using to index it. So when we create the first part here,
we're setting the `exact_path` to
https://www.gov.uk/foreign-travel-advice/afghanistan/warnings-and-insurance.
This path does resolve, but only after a redirect to the base path.

So to be consistent with how the GOV.UK website works, we should omit
the slug on the first part.

[1] https://www.gov.uk/foreign-travel-advice/afghanistan