Add settings audit model and migration
This adds a very basic model which will be used to store some very basic
audit information about the settings model.

I've gone with a specific model name that couples it to the settings
rather than a generic audit model as this does seem pretty coupled with
it's intention/implementation.

I've also had to somewhat fight against my instinct to use a foreign key
for user to ensure this models as simple as possible. It might be that
it's better to use a has_one relationship rather than storing an id and
email address on the audit itself if we want to expand the functionality
of the audit class in the future.