Replace hashie for managing prompt structures
Using hashie for objects as nested as our prompts was causing annoying
warnings:

```
W, [2024-07-31T08:30:18.945090 #39224]  WARN -- : You are setting a key that conflicts with a built-in method Hashie::Mash#default defined in Hash. This can cause unexpected behavior when accessing the key as a property. You can still access the key via the #[] method.
```

We concluded that we had probably pushed Hashie further than we should
have as we're not navigating through these config options as objects and
have a number of keys we cannot change (as they are the structure of a
JSON schema).

Instead we now use ActiveSupport::OrderedOptions for the config option.
This provides us the means to look up the first level of options with a
method and then we need to traverse the items as hashes after that. This
seems sufficient for our prompts where the top level key is quite a
meaningful value but after that feels are more arbitrary.

I used with_indifferent_access on the hashes so that we could look up
the values in them via symbol or string, as using symbols feels more
idiomatic in the code.

Doing this was a little challenging combined with
ActiveSupport::OrderedOptions, as that class will only provide methods
for a class with actual symbol keys and will not provide them for a
HashWithIndifferentAccess, hence why we have to do the symbolize keys
and then transform_values.