Change feature name in spec
This spec just asserts a bunch of stuff around feature flags.

The whole feature flag implementation was added when we wanted to test a
new answer strategy, and so the name of the feature used in the test was
the name of that first feature.

Really all we care about in this test is that the feature mechanism is
working, so we can just use a fake name for the feature instead. It
removes any confusion about what the test is doing. It also removes
any reference to "chat_api" from the codebase, which is old answer
strategy we were using but aren't any more.