Update ruby version in Gemfile to use pessimistic version constraint
We recently encountered an issue where the GOV.UK Ruby Images were
bumped to 3.3.4
https://github.com/alphagov/govuk-ruby-images/pull/106.

We were running our app on 3.3.3. This caused our deploys to integration
to fail due to there being a Ruby version mismatch during the image build.

Kevin found this solution to the issue. By ussing pessimistic version constraint
(~>) in the Gemfile, we can avoid this issue in the future.

This will allow patches to be applied up to 3.4.0 without us having
to manually bump Ruby each time an upgrade is released for the Ruby image
in the govuk-ruby-images repo.

We will still need to manually bump minor versions though.