Restore Heroku release task
As per [1] we had problems with setting this task to run on a non-eco
dyno. Now that we're in a new month with new free hours we can run this
again.

I'm still waiting on Heroku support to explain how to change the dyno
type for a release task on review apps.

[1]: https://github.com/alphagov/govuk-chat/commit/d1e0d5f973c9d2bc46f8c69991280c37c9ffcd0b