Add LinkTokenMapper to convert links to tokens
This adds a new `AnswerComposition::LinkTokenMapper` class which exposes
one method for taking in some HTML content and rewriting all of the
link hrefs to be a token value instead of the actual link.

It uses Nokogiri to traverse the DOM and pick out any `a` nodes. For
each one it finds, it rewrites the `href` attribute to be "link_{i}"
where `i` is the next positive integer after the current stored count.

It also stores each link it rewrites into its `mapping` variable, which
is a hash mapping the token to the original href.

Duplicate links will be given the same token, i.e. the token is
retrieved and set, rather than a new one generated.