Rewrite HTML links in answer composition
This updates both the structured and unstructured answer composers to
rewrite URLs before including them in the system prompt.

This is to try to solve the problem of the LLM hallucinating and making
up its own links when we tell it what sources to use.

The idea is that if we switch the href of a URL we give it to a token
instead, it'll return us the token in the response and we can substitute
the real value back in.

For example if we give the LLM some context (i.e. some chunks of HTML
from OpenSearch) and those chunks include links to other GOV.UK pages,
we don't want it to ever change those URLs for any reason, otherwise
they'll point to broken pages.

So if a chunk looks like this:

```
<p>Click here to <a href="/find-training-courses">find a course to apply for</a></p>
```

Then we'll rewrite it to:

```
<p>Click here to <a href="link_1">find a course to apply for</a></p>
```

When rewriting, we store the mapping in memory so we can reverse the
substitution later on.