Add method to replace markdown tokens with links
This adds a new method to the `LinkTokenMapper` which takes in some
markdown and rewrites any token-based URLs (e.g. "link_1") with the
value stored in the mapping.

If the mapping doesn't exist, we strip that link from the markdown and
replace it with the link's content instead. This would usually just be
the text content, but it can be a more complex markdown tree as well.