Update answer composers to rewrite token URLs
When we send a request to OpenAI, we're rewriting all the URLs which we
give it from the relevant chunks we retrieved from OpenSearch.

So a URL will have been changed from, e.g. "/check-state-pension" to
"link_1". The idea is that the LLM will include these links if it needs
to, but it won't be able to hallucinate and alter them.

When we get the response, we need to then reverse this rewriting so that
the URLs make sense again.

This commit updates both the structured and unstructured answer
composers so they do that reversal process.