Add Settings::BaseForm
This commit adds a base form for settings forms. The purpose of this form
is to setup all the functionality required for other settings forms which
will handle setting the correct value and validating the attribute it
is intended to set.

The reason for adding this base form is there's a tonne of functionality
that every settings form will need to implement. These are:

- including the ActiveModel::Model and ActiveModel::Attributes modules
- adding the user and author_comment attributes
- validating the shared author_comment attribute
- loading the settings singleton instance

This form will be inherited by all other settings forms. They will:

- define the attribute they are updating on the settings instance
- validate the attribute they are updating
- implement the logic required to set the attribute on the settings instance
- call #locked_audited_update on the settings instance with the args
required to create an audit and set the attribute on the settings
instance within the passed block.