Add settings show page to admin UI
In a previous PR we added the Settings singleton which will be used to
control access. This commit adds a show page to the admin UI which will
be used to view and eventually link to edit pages for each setting.

It's a pretty basic page and we might want to redesign it later, but for
now I think using summary list components is fine.