Add endpoint for updating instant access places
This adds an endpoints for editing and updating instant access places.
It uses the new forms added in the previous 2 commits.

It's pretty straightforward. I've chosen to add a new controller for each
setting as discussed in the spike PR. We may want to rethink that down the
line if we end up with tonnes of settings.