Map links to tokens in search results
When we perform the search against the OpenSearch index, we get a bunch
of results back. Each result contains a `page_url` attribute, which is
ultimately how we figure out whether the LLM used that particular page
when coming up with its answer.

In order to reduce hallucinations, we want to tokenise these links, i.e.
substitute them "token_{n}" before putting them in the prompt, then
reversing that process when we get the response.