Remap URLs before setting context source
The response we get from the LLM in the `sources_usd` part of the
response will contain links in a tokenised format, rather than actual
URLs.

For context, we're not providing the LLM with any actual URLs any more.
Instead, we substitute each URL with a token when sending the request,
then convert the tokens back to URLs when we get the response.

This means that the LLM will be telling us it used "token_1" as a source
URL. We need to reverse the mapping back to real URLs before we can
figure out which sources were actually used.