Add Settings::InstantAccessPlacesForm
This commit adds the form that will be used to edit the instant access places.
There's a few things of note in this form that might be slightly strange
at first glance:

If a user inputs a negative number of places and the total of that
number subtracted from the total number of places is less than 0 then
we would end up with a negative number of places in the db.

To ensure that happends we doesn't happen this:

1. adds a validation to ensure negative numbers cannot be persisted
2. adds a failsafe which sets the number of places to 0 in the db if
for whatever reason validate! returns true

Since the instance is locked prior to update hopefully this won't be required.

Due to the above the action passed to the audit needs to reflect
the actual number of places removed if the failsafe is triggered. So there's
a bit of logic in the "#action" private method to handle that.