Add sign up enabled form
This adds the form object used to update the sign up enabled setting
and creating a setting audit.

I figured we don't want to create an audit if the setting hasn't changed
and the user hasn't left an author comment.

It might be better to just return even if there is an author comment though.