Change technique for delaying question loading
The previous technique no longer works because we don't use
Question.create! anymore in the controller and instead build a new one
with `new` and then save it with `save!`.

At first, I tried putting a sleep in on a stub of the `new` method -
however this created a very slow infinite loop when building a question
within it (which must also use the new method). I decided it was better
to stub new to return an object and put the sleep on a method call of
that - at least any problems would be revealed faster.