Set a default browser title
This sets the browser title to be GOV.UK Chat or have a suffix of it
whenever it is set. This prevents there being an empty title.

At the moment we don't have a means to set different browser titles and
page titles and I'm not aware of a need (yet). We'll have to configure
content_for different to allow this.