Set titles for onboarding/conversations
This sets titles for onboarding limitations, onboarding privacy and a
conversation.

These have been done in the unconventional way of setting them in an
instance variable on the controller. The reason for this is so that this
same variable can be used in both HTML rendering and JSON rendering,
this avoids the need to duplicate the title.

I took the titles from:
https://github.com/alphagov/govuk-chat/pull/478/files however I did
change the one for a conversation from: "GOV.UK Chat (main page)" to
"Your conversation" as that seems weird and just didn't make sense to me.
I did feel the other titles are maybe a bit unconventionally wordy for
browser titles but didn't seem like they were committing any crimes.