Update page and browser titles for conversation by JS
This builds on the previous commit which exposed the data of these title
values via the JSON responses. This code takes that JSON response and
uses that to set the document.title and the h1 title values.

The code is set-up to be forgiving if the browser title lacks " -
GOV.UK Chat" or has additional dashes within it.