Ensure conversation inline links are absolute
As we don't host GOV.UK Chat on www.gov.uk at the moment, any relative
inline links in content don't work. Even once we are on www.gov.uk we'd
still have the admin area and heroku environments where these link
wouldn't work.

To resolve this this commit adds the functionality to convert all the links
to be absolute links at the point of answer generation.

To do this required dealing with a few dilemmas:

1) Where to put this code? I decided to put this code inside the
   LinkTokenMapper as we are already traversing the markdown of an
   answer and I felt it'd be good to avoid having to traverse this code
   twice - and particularly not at render time. I know this is a bit
   janky so happy to chat over alternative approaches.
2) Kramdown will write inline absolute links with references and won't
   use the more common bracket approach [1] - this makes all the tests
   related to links harder to write as you need to consider the link
   references in the footer.

[1]: https://github.com/gettalong/kramdown/blob/bd678ecb59f70778fdb3b08bdcd39e2ab7379b45/lib/kramdown/converter/kramdown.rb#L294-L301