Use Plek.website_root for evaluation code
This code was inconsistent with the runtime code which used
Plek.website_root to create absolute URLs for sources. This was done for
good reason, it's unlikely you'd want these sources to not be references
to https://www.gov.uk, but it is now inconsistent with answer code.

The answer code now also uses Plek.website_root for inline links which
leads these links to then be inconsistent with inline links.

By using the same Plek.website_root method we can keep things
consistent.