Add secondary_navigation component
We need a mechanism to navigate between the settings show page and
settings audits index page. In Content Publisher and Whitehall we
used a secondary navigation component to achieve this.

This commit adds the component based off of this PR
https://github.com/alphagov/whitehall/pull/6975.

Which in turn was based on the Content Publisher implementation
here https://github.com/alphagov/content-publisher/pull/1576.

We should enquire about moving this to the Publishing Components gem
since a few apps use it, but for now since we're pressed for time i
think it's best to just copy the component over.