Set session cookie to expire after 30 days
This has been done to simplify decisions that we are making related to
data stored in session and when it will expire.

This changes the session cookie from being something that expires at the
end of the browser session - a hugely ambiguous term for modern web - to
be something that expires after 30 days. I had to manually specify the
key as without this argument we end up with a key of "_session_id" the
"_govuk_chat_session" reflects what this key previously was.

It's worth noting that each time a new request is made the 30 days
resets, so a session can be effectively infinite.

I chose 30 days as it matches what we communicate to users about how
long they can access conversation and stay signed in with a magic link.