Add pagination helper to build previous_and_next_page_hash
There's a lot of duplication between the pagination of questions and
audits.

This pulls the common logic out into a helper method so that we can
reuse it in both places and any others going forward.

One slightly awkward thing with is that the next page url and prev
page url need to be passed in so that we can ensure we retain all
query params for questions.