Set an environment variable to disable early access authentication
This sets up a rather contrived conditional to reflect that the homepage
will need to be used in two circumstances: early access on (production)
and early access off (heroku demos).

My hope is that by setting this up early we're able to consider it while
we're implementing early access as it would be easy to forget.

I also updated the existing one for signon to be defined in the same way
that will produce a bool. I expect that env var will be removed once we
progress further with early access.