Add a revoked_at column to early_access_users table
This column is to be used as a way to prevent users from having further
access without deleting their account - effectively as a banning
measure. If this field as a value a users access is considered revoked.

I expect we may also want to record some audit information with this,
such as a comment, when implementing the controls for this feature.