Add Question Router class
This adds the pipeline class for performing question routing.

It makes a call to the LLM which uses function calling to classify a
user's question. It'll return a label which we'll store on the Answer
model along with some other attributes.

If the label is `genuine_rag` then we'll continue with the pipeline to
the next step (currently search result fetching). If it's any other
label, we'll abort the pipeline because the question isn't relevant for
us to answer.

The classifications are all stored in a config file and the OpenAI
request body is built up using this data.