Add basic step by step nav parser that handles paragraphs
We're going to start indexing step by step navigation documents into
our OpenSearch index. This commit adds a basic parser that can handles
parsing the step_by_step_nav hash. See here
https://github.com/alphagov/publishing-api/blob/5c80d0bd336152c35815bad05d912dcd9f31f5e1/content_schemas/formats/shared/definitions/_step_by_step_nav.jsonnet#L25-L112

Step by step nav documents differ from other content types we've added
in that the details hash provides a list of steps, a title and an
introduction.

Each of these steps has a title and an array of content blocks. These blocks can
be:

- a paragraph
- an ordered list of links
- an unordered list of links (bulleted list)

An unordered list of links is indicated by a key value pair
on the list content of { type: "list", "style": "choice" }.

It also has to be able to handle an additional key on the the step of
```
{ "logic": "and" }
```
or
```
{ "logic": "or" }
```

When this is the case it means that the step follows on from the step
before and should either be done in addition to it or instead of it.

I'll handle parsing lists and applying the 'and' and 'or' logic in
follow up commits.

This commit adds a basic parser that can handle paragraphs. For now it
only adds functionality to:

- add the introduction to as a p tag at the start of the html
- add the step title as an h2 tag
- parses paragraphs for each step and adds them as p tags