Add logic for parsing ordered and unordered lists
This builds on the previous commit by extending the #parse_nav_step method
to handle parsing ordered and unordered lists and converting them to HTML.

It determines if the list is ordered or unordered by checking the for
a 'style' key value pair on the list content block.

If the 'content["style"] == "choice"' then it will construct the list as
an unordered list, otherwise it will construct the list as an ordered list.

Each list contains 1 or more link items which are formatted as follows:

{
  "href": "/path/to/step",
  "text": "Step 1",
  "context": "Optional context for the link. This is rendered in a span tag."
}

Each link item is wrapped in a list tag and uses the 'href' and 'text' key
value pairs to construct the link. There is also an optional 'context' key
value pair which is rendered in a span tag after the link.