Handle 'and' and 'or' logic in the step by step nav parser
Each step is rendered on the frontend as a numbered step, an 'or'
or an 'and'.

Any step can be followed by an 'or' or an 'and' step. And's should
be done in addition to the numbered step. Or's can be done instead.

The logic key value pair that we can use to determine this sits on the
top level of the step hash:

And
```
{
  "step" =>
  {
    "title" => "Step 1",
    "logic" => "and",
    "contents" => [
      {
        "type" => "paragraph",
        "content" => "This is an 'and' step."
      }
    ]
  }
}
```
Or
```
{
  "step" =>
  {
    "title" => "Step 2",
    "logic" => "or",
    "contents" => [
      {
        "type" => "paragraph",
        "content" => "This is an 'or' step."
      }
    ]
  }
}
```

This commit adds logic to the step by step nav parser to handle this by
checking if the next step is an 'or' or an 'and' and if so adding an
additional paragraph to the step.

This paragraph is added to the HTML prior to the rest of the HTML like so:

```
<h2>Step 1</p>
<p>This is a numbered step.</p>

<p>and</p>

<h2>Step 2</h2>
<p>This step has 'and' logic.</p>

<p>or</p>

<h2>Step 2</h2>
<p>This step has 'or' logic.</p>
```