Use FactoryBot build strategy on authenticatable association
This is a follow up to a comment thread in [1]. Where we were required
to a DB write to build a passwordless_session. The regular technique of
`authenticatable { build :early_access_user }` wasn't working,
presumably because auto_save is off [2].

Using the association name resolves this and has the early_access_user
utilise same creation strategy as the passwordless_session factory.

[1]: https://github.com/alphagov/govuk-chat/pull/468#discussion_r1709746987
[2]: https://github.com/mikker/passwordless/blob/435f3df152f60e98c97bbcf95cb77d80d98641a2/app/models/passwordless/session.rb#L13