Ensure content passed to html is escaped
On the off chance that we recieve a malicious payload, we should ensure
that we escape any HTML in within structured content prior to saving it
to Opensearch.

This updates the parser to use the .tag method rather than manually
writing the HTML so we don't have to escape it.