Update question rephrasing prompt
The data science team have come up with a new prompt for question
rephrasing.

It's a little more complex than the one we had before. Previously we had
the system prompt followed by the last 5 messages in the conversation
formatted as `user` and `assistant` blocks.

Now we have a system prompt followed by a single user prompt. The user
prompt contains:

* a preamble
* the message history, but this time formatted as a newline separated
  string
* an instruction
* the current question
* an outro

The logic for retrieving the conversation history needed to be tweaked
slightly here. Previously we could just grab the last 5 question/answer
combos, because the current question would be the last one.

But now we want to separate out the current question in order to embed
it in the last part of the prompt. So the AR query needs to specifically
exclude the current question.