Re-introduce the concept of true/false values
This is in order to determine true_positive?, false_positive? etc to calculate
precision/recall.

This is done using a lambda to determine true/false status.
The reason for this is we want to have exactly the same metrics for the
classification approach to guardrails so we can compare. They have a slightly
different way to evaluate truth, otherwise the evaluation will be identical.