Use pessimistic locking to prevent concurrent magic link access
This adds a lock so that we get two requests at the same time with the
same magic link, the second request will have to wait for the first
request to have completed.

It's very difficult to test the actual scenario, as it requires spawning
additional threads, so I've gone for the basic assertion of lock being
called.