Redirect users to conversation rather than chat_page
As we intend the chat page to be the sign-up point, this isn't the ideal
spot to redirect.

Instead of chosen onboarding_limitations_path for a sign-in, with an
expectation that might be their first sign-in, if they've already got a
conversation they'll be redirected again, and show_conversation_path for
an already signed in user, with an expectation they may already be
chatting - again they'll be redirected again if not.

I did initially plan to sign a user out if they tried to access a magic
link a second time. I had an expectation that this would allow handling
a few edge case scenarios more elegantly, such as comms if re-using an
old link or if you're trying to change user account - but thinking about
it I think just keeping them logged in is simplest and will at least
mean that visiting the same magic link again - if you've remained logged
in - doesn't kick you out.