Create basic views for magic link failures
This makes use of draft content from the content designers, so may be
tweaked further - and I've also been a bit creative.

I decided we can handle pretty much all error scenarios with 2 views
(could even be 1 I guess) rather than as many as were being designed
for.

I decided that it was probably most appropriate to show the invalid or
expired link whether or not the link was ever valid or not. This has the
benefit that any links we delete [1] don't give users a confusing
response. This comes at the expense of showing this message for
obviously wrong links like /chat/sign-in/a/b. Should this not be
acceptable we may want to auto-expire links rather than delete them and
then use a regular 404 page.

[1]: https://github.com/alphagov/govuk-chat/blob/bca4f02441018741b72167175e3a2bce0771d46f/app/models/early_access_user.rb#L17-L20