Use rephrased question in rephraser message history
This was a request from DS.

When we rephrase the user's question, we build up a message history
comprising the last 5 user<->assistant interactions, then tell the LLM
to rephrase the user's question if necessary.

To improve the accuracy of the LLM here, we should be including the
_rephased_ user's question in the message history, not the question they
actually asked.

If a user asks something vague, for example, and the LLM has rephrased
that question previously to turn it into something concrete and
understandable, then that's what should be included in the message
history on future rephrasings.