Add DB columns for being able to turn off public side of application
This applies the initial groundwork for using the admin area to turn off
access to chat to display a downtime page.

The boolean, `public_access_enabled` field determines whether public
access is on or off.

The enum, `downtime_type` field determines whether the application is
turned off temporarily, and thus returns a 503 response, or permanently,
and thus returns a 409 response