Split HTML sanitisation from chunking
This refactors the HtmlHierachicalChunker to have its sanitisation code
extracted to a separate class. The motivation for this change is that we
have found a distinct need for the sanitisation code without the
chunking in [1].

In refactoring this I saw opportunities to reduce the quantity of
testing as there were a number of tests that didn't seem to test logic
that was coded.

The refactoring also provided an opportunity to remove the
AttributeStripper class as, in the new lower responsibility
HtmlSanitiser class, this didn't seem a necessary abstraction.

[1]: https://github.com/alphagov/govuk-chat/pull/502