Update test
This test was failing as it was looking for a button that only had a class on it and no data attributes, which have since been added for GA4 tracking. I could have added the data attributes to the test but I noticed that we take a bit of a lighter touch in an earlier limitations test ('when the request format is JSON > when session[:more_information is true] > it returns a successful response with the correct JSON'), where we just test that form_html contains the text (i.e. /Take me to GOV.UK/) rather than testing that the text is also inside a link. Therefore, I mimicked that approach and modified the test in question to just check against /Okay, start chatting/.

Happy to discuss.