Configure BaseController for settings.public_access_enabled
This sets up a before_action that will intercept requests and check
whether settings.public_access_enabled is false. If so, it will return a
response to explain the application is unavailable. This only applies to
the frontend routes of the application and not the admin area.

The status code returned will depend on the settings.downtime_type
variable, with 410 if the application is permanently offline and 503 if
temporarily offline.