Admin settings form for public access
This adds a form for the admin settings area which provided the control
to turn public access to chat on and off and, for when it is off,
whether the downtime status code is temporary (503) or permanent (410).

Unlike the other settings forms this contains two fields.

I did wonder if the system tests should actually test the side effects
of the change, perhaps this is something we should consider when we have
more of the settings features wired in?