Adapt tests that use ActiveJob for Sidekiq queue adapter
In Rails 7.2 they have changed the behaviour of tests that make use of
active job to use the active_job.queue_adapter value which means that
the methods `perform_enqueued_jobs` and `enqueued_jobs` are no longer
available and instead methods from the queue adapter need to be used.

I thought it was better to adapt the tests to use Sidekiq rather than
make the tests less representative of reality by using the test adapter,
it does however make them a bit uglier.