Add user description form
This will handle the first question we ask users when they sign up for
instant access.

Since we don't actually do anything other than store the answer in the
session i've not added a submit method to the form.

I can see an argument for adding one and returning a Result object
so that we follow the same pattern other forms.